patch.h: initial revision, patch256
So let's try it.  We want to compress 16-bit sorted integers, for use
with pkgelfsym-data/topsym.list.  We don't want entropy coding, which
is a bit too slow (I tried FSE, it takes >10 cycles per integer while
compressing down to ~4 bits).  We don't aim for best compression but
rather for fast decompression: say, 3 cycles, down to 6 bits.

This is doable if we compress deltas.  A lot of deltas will be very
small, because topsym.list is sorted and the references are correlated,
e.g. a dozen of SDL_* symbols in a row, followed by a gap, followed by
a few png_* symbols, etc.

This is ideal for the PFor-Delta compression scheme (and its variants
NewPFD and OptPFD).  Decompression works in 3 stages:
- Decoding a block of m-bit integers, much like with Golomb-Rice code.
- Extending some of the integers with extra high bits.  This is also
  called "patching the exceptions".
- Summing the deltas (possibly over a few blocks).

Only a few deltas need to be patched, roughly 1/8 of them.  But it can
only be done with a scalar loop, while other stages are SIMD-amenable,
doing 8 integers at a time (or 16 with AVX2).  Thus patching can easily
become the most expensive stage.

Therefore I adopt a very simple, bytewise patching specification: the
number of exceptions, followed by <index, high bits> tuples.

    +---+ +---+---+ +---+---+
    | n | | i | x | | i | x | ...
    +---+ +---+---+ +---+---+

Note that for blocks of 256 integers, there is no need to check if the
index is out of bounds.  There is in fact only one condition in the
loop, which is the loop control!

I further try to unroll the loop in a clever way:
- The number of exception cannot be 0 (this can be signalled by a
  separate flag instead of wasting a byte).  Yet I want to avoid
  a separate check for n == 0.
- Unrolling by a factor of 2 can be done with little extra branching
  (recently I did something similar in rpmset/downsample.c).
- The compiler should transform the while loop into an if + do-while,
  so there is already a branch to take advantage of.

With n=32, the loop runs at about 3 cycles per integer.