patch.h: fuse tuples with the same index into 16 bits
The problem is that extra 8 high bits might not be enough (with m < 8).
There are a few ways to solve the problem:

1) Using vbyte encoding:

    unsigned i = s[0];
    unsigned x = s[1];
    if (unlikely(x & 0x80)) {
	x &= 0x7f;
	x |= s[2] << 7;
	s++;
    }

This is enough with m > 0.  Note however that there is an extra branch
per each integer.  Moreover, because the encoding is variable-length,
a single boundary check may not be enough - we then need a separate
version of this loop which does boundary checking on each iteration.

2) Running the second loop:

    patch256(s, n, v, m);
    s += ..., n = ...;
    patch256(s, n, v, m + 8);

A separate loop can be avoided if the values have to be extended only
by one bit (e.g. this is always the case with m=7): if we ADD high bits
instead of ORing them, we can supply two tuples with the same index.

3) A technique that I invented this morning: we can fuse two tuples
with the same index to obtain a full 16-bit addend.  Tuples can only
be fused at an even offset, so this takes one extra branch per two
integers.  We can further avoid branch misprediction if a value needs
to be extended only by 1 bit, by placing the tuples in such a way as to
elude fusion.

This new loop actually runs quite a bit faster on x86, and not much
slower on arm.  I have no idea why the loop with an extra branch can
run faster, my guess is that by factoring a common subexpression of
(x0 << m) the compiler is able to produce a better instruction
sequence.  So there's no need for different patch256 versions for now.