bitpack16.h bitunpack16.h: initial revision, 5-bit packing
These files contain macros which describe how to pack and unpack
5-bit integers in abstract terms, as if using only one 16-bit lane.
We generate the packing and unpacking routines by cross-breeding
the macros with SIMD engines (in C++ terms, the engine is a template
parameter).  For example, to obtain bitpack16_5x64 and bitunpack16_5x64
routines, we combine BitPack16_5x8 and BitUnpack16_5x8 macros with
the B16x8 SIMD engine.

The smallest size we are able to go down is 5x8, by combining 5x4 with
B16x2.  This size is the smallest possible (40 bits = 5 bytes).  Note
that the 5x4 macro uses quarter-load, unlike the 5x8 macro which
requires only half-load.  This means that 5x4 cannot be combined with
B16x8 and B16x16 which do not support quarter-loads.