bitpack16-simd.h: quarter-load for B16x8
It occurred to me that even within the SSE2 instruction set,
quarter-loads can be implemented relatively efficiently.  We want
to load 4 bytes and distribute them among 8 lanes, 4 bits per lane.
The order of the 4-bit segments is:

	      |7531|6420|
	          ^    ^
I mark the segments that are already in the right place.
We then shift left by 28:

    |    |    |7531|6420|
      753 1642 0  ^    ^
        ^    ^
With another such shift, all segments are in the right place.
The trick is that the second shift is to a byte boundary.

Granted, this is not as efficient as pmovzxbw for half-loads, but the
comparison is beside the point.  We should rather be comparing how B16x8
with quarter-load fares against B16x4 with half-load, implemented with
shifts and masks as well.  And it fares favorably.

On Haswell:
-unpack16_5x32: 0.595 cycles per integer
+unpack16_5x32: 0.313 cycles per integer

On Ryzen, compiled with -mavx2:
-unpack16_5x32: 0.496 cycles per integer
+unpack16_5x32: 0.321 cycles per integer

As expected, the difference is more pronounced on Haswell.  (In an early
version of bitpack16-simd.h, I had two B64x4 implementations: one backed
by uint64_t, and another by XMM half-register.  The latter worked better
on Haswell while the former on Ryzen.  Haswell's SIMD engine, limited to
64 bits, performs better than its 64-bit integer engine; alternatively,
Ryzen's integer engine is faster because of its extra execution units
and its 6-wide uop dispatch.)