enc64.c: initial revision, rdopt64 rdopt128 rdopt256
So let's try regular blocks with 64, 128, or 256 integers, and the
trailing block with less than 64 integers.  There are two problems:

- How to best encode a single block of 64, 128, or 256 integers.
- How to best tile the input with blocks of 64, 128, or 256 integers.

This commit addresses the first problem.  To encode a block of
64/128/256 integers, we try different values of m to minimize the
total length, which is the length of SIMD block + the length of patch.
Note however that longer patches take more time to apply during
decoding.  Let's consider the worst case of how expensive a patch
can be.  If we have a block of 256 integers and move from m0 to m0-1,
this save us 256 bits, which is 32 bytes.  Those 32 bytes can hold 16
patch exceptions.  The worst case is when the length is reduced by 2
bytes, but the number of exceptions is increased by 15, taking about
2.2*15=33 more cycles during decoding, increasing the cost of decoding
by about 0.13 cycles per integer, or by up to 2x.  Another way to look
at this is that we trade bytes for cycles, but this particular trade
deal didn't go well: each of the two byte saved costs us 16.5 extra
cycles on decoding.

The analogy of trading bytes for cycles is insightful, but its naive
implementation can be suboptimal: we should also take into account the
decoding cost of SIMD block.  E.g. the cost decreases when going from
m=9 to m=8, and increases again when going from m=8 to m=7 (because m=8
is such a cute number).  We may condone a longer patch in the first case
but not in the second.  Therefore, instead of devising a heuristic which
limits the patch length, we invoke the general theory of rate-distortion
optimization.

In lossy compression, we want to minimize the bitrate, which is the
"rate", while also minimizing lossy artifacts, which is the "distortion".
Or, in our case, the "distortion" that we want to minimize is the
decoding cost.  The right way to do this, as summarized by Fabian
Giesen [1], is to choose among all solutions such a point <len,cost>
that minimizes

     LenRD * len + cost,

where LenRD scales the length units against the decoding cost units and
also tells how important optimizing the length is, relative to optimizing
the decoding cost.  In the code below, I refer to this RD-weighted combo
as "TCO", the total cost of ownership.  So the solution to the fist
problem is to minimize the TCO of a given block.

[1] https://fgiesen.wordpress.com/2018/12/10/rate-distortion-optimization/