enc64.c: optimal tiling
The answer to the second question, how to best tile the input with
blocks of 64, 128, or 256 integers, is: tile in such a way that
minimizes the sum of the TCOs of the blocks.  We then need a technique
to evaluate various combinations of 64+64 and 128+128 blocks.  Note that
it is not enough to e.g. split a block of 128 integers into 64+64 and
choose between 128 and 64+64.  This way a block of 128 integers can only
end up on a 128-integer boundary.  Somehow we need to try and recombine
the blocks "in a window".

Therefore we resort to a more general technique reminiscent of LZMA-style
forward parse.  For each block of 64 integers, we calculate the best way
to encode the input segment up to and including the block.  In relation
to optimal parsing, this is called the "arrival cost".  The inductive
step is: by knowing the arrival cost at (i-1), and further at (i-2) and
(i-4), we can deduce the arrival cost at i (indeed, by trying to combine
the block with the previous blocks).  This technique seems to find the
optimal cost in linear time and with no extra work.