the new patching scheme
Fused entries are wasteful: it takes 4 bytes to clobber a 2-byte value.
They make sense from the generality point of view (arbitrary inputs are
encodable with arbitrary m), but we need them relatively rarely in
practice.  Now I limit them to one 3-byte clobber entry per block.

We want then to reduce the patch size even further.  One option is to
delta-encode the indexes.  For patch256, we can use 7-bit index deltas.
But there is no guarantee that all the index deltas fit into 7 bits,
and in practice we do have big gaps.  And then it takes another branch
to decide whether indexes are 8-bit values or 7-bit deltas.

Another option is to cut down on the values.  It took me a couple of
days to realize that this is actually a better option.  In the same way
that roughly 90% values fit into m bits, we should reasonably expect
many values on the patch list to be small.  So if we manage to encode
two values in a single byte, we save one byte per two (i,x) tuples,
which is much better than saving 1 bit per tuple with 7-bit indexes.

Another problem with patch256 is that 8-bit indexes are not memory-safe
with x64 and x128 blocks.  Yet I don't want to check each index for
(i < 64) resp. (i < 128).  The best way is then to limit indexes to 6
resp. 7 bits, and redistribute extra bits to values.

So now we've got two kind of patch entries:

    x64:
    - 3-byte combined: (6,6)+(6,6)
    - 2-byte single:   (6,10)
    x128:
    - 3-byte combined: (7,5)+(7,5)
    - 2-byte single:   (7,9)
    x256:
    - 3-byte combined: (8,4)+(8,4)
    - 2-byte single:   (8,8)

The downside is that now we need to run two patch loops, but this is not
that bad compared to what we already have.  The patch256 loop is already
unrolled:

    do {
	x0 = s[0];
	x1 = s[1];
	i0 = s[2];
	i1 = s[3];
	s += 4;
	v[i0] += x0 << m;
	v[i1] += x1 << m;
    } while (s < last);
    if (s == last) {
	x0 = s[0];
	i0 = s[1];
	v[i0] += x0 << m;
    }

We replace it with two loops.  The first loop, the one which deals with
the combined 3-byte entries, still has to be unrolled, but the second
loop will handle fewer values (on average) and there is no need to
unroll it.  Compared to the odd case (s == last), the second loop is
only moderately more expensive (one branch to enter the loop, and one
branch misprediction at the end of the loop).

As a result, patching becomes measurably slower (1.02->1.09), but we've
saved a good deal of bytes (1461->1377).  The trade-off seems favorable.

LenRD=16: len=1451 dec:1.03 cycles per integer
LenRD=32: len=1377 dec:1.09 cycles per integer
LenRD=64: len=1357 dec:1.15 cycles per integer

Another problem is that the number of iterations is now limited by two
4-bit fields, so we can patch at most e0=15*2 + e1=15 entries.  This
limitation can be noticeable with x256, it would be nice to have e0=40
and e1=20+.  Maybe we can procure another bit from bctl, and split 8+1
bits into something like 22*23 < 512.