trying to extend x256 e0=0..46 e1=0..31
Update 2021: I've just reviewed the scheme and it makes sense.  We have
roughly (m=0..16)x(N=64,128,25) > 48 distinct types of blocks as indicated
by bctl (the leading byte).  The best we can do to dispatch the cases is
to switch on 6 bits of bctl.  The compiler will generate a jump table
which is roughly 64*4 bytes or 4 cache lines.  (If we switched on every
possible bctl value, we would squander 16 cache lines.)

Hence we switch on 6 MSB bits while 2 LSB bits convey extra information:

    [765432|1|0]
          ^ ^ ^
	  | | |
	  | | `- has pctl + patch
	  | `--- has a clobber entry / outstanding tuple
          `----- block type (m=0..16)x(N=64,128,256)

The whole block looks like this:

    [bctl] [blob] [i|w] [pctl] [patch]
               ^     ^              ^
	       |     |              |
	       |     |              `- packed patch tuples
               |     `---------------- clobber entry / outstanding tuple
               `---------------------- packed m-bit integers

The pctl byte specifies two nibbles: e0 for the unrolled loop with
small/paired MSB quantities, and e1 bigger MSB quantities.  But the
nibble values are limited to 15, which is not quite enough for N=256.
Therefore, for some N=256 blocks we allocate two cases in the dispatch:

    case C + 0:
    case C + 1:
	    bitunpack16_##M##x256(v, src);

so that we can take advantage of bit #2.  LSB bits in bctl are then
interpreted differently:

    [76543|2|10]
           ^  ^
	   |  |
	   |  `- has pctl + patch
	   `---- has a clobber entry / outstanding tuple

If (bctl & 3) == 0, there is no pctl + patch.  Otherwise, we have pctl +
patch and extra 3 possible values: 0b01, 0b10, or 0b11.  We use them to
extend the pctl byte: the range of e0 is now 0..23(x2), and e1=0..31.

When m >= 8 (with N=256), there is no need for a 16-bit clobber entry:
an 8-bit entry from e1 is enough to make up for any value.  Hence we can
reuse bit #1 to provide extra state, without allocating two cases in the
dispatch:

    [765432|10]
             ^
	     |
	     `- has pctl + patch

With N=64 and N=128, we do not extend the ranges of e0 and e1.
Hence with m >= 6 resp. m >= 7, bit #1 remains unused.

Finally, when m >= 12 (with N=256), 4-bit paired quantities are always
enough to make up for any value, so we pass e1=0 and expect the compiler
to optimize out the e1 loop.  The only gotcha is that the entries are
paired, but we may have an odd number of them.  So we reuse the
outstanding entry (also called the "f tuple", not sure why): bit #1
indicates that the odd entry is encoded with 2 bytes (unlike "full
clobber" entries which are 3 bytes).