enc64.c: patch cost more accurate
Based on the previous commit, the cost of e0 loop is now assumed to be
5.4 cycles per iteration (N=64,128) or 4.8 cycles per iteration (N=256);
e1 respectively 2.5c or 2.0c.

This also introduces more realistic costs for branch misprediction 15c,
and block tax / indirect jump 20c.  As a result, the encoder is now
more bent towards N=256 blocks.

-len=1310 dec:1.02 cycles per integer
+len=1387 dec:0.93 cycles per integer

Let's see if the change is favorable.  We can save 1387-1310=77 bytes,
but it would cost us about (1.02-0.93)*2173=196 extra cycles on
decoding, or 196/77=2.55 cycles per byte.  Which is indeed close
to the threshold LenRD=3.2 cycles per byte.