enc64.c: fixed ffind and f's handling in encpatch
The f entry (the clobber entry / outstanding tuple) wasn't handled
properly in encpatch.  When encoding a patch, the f entry should be
skipped.  It's easier to pass the index of the entry rather than
recognize it by the value.  I don't quite like the resulting code,
but finally it seems to work.  This is much better than broken code!
I'm now able to encode and decode all integer lists associated with
pkgelfsym-data/topsym.list.