dzag16enc.c dzag16dec.c: initial revision, delta + zigzag
When a list is not sorted, there will be negative deltas (with the
sign bit and typically a few more MSB bits set, due to two's complement
representation).  The idea of zigzag encoding is to store the sign bit
in bit #0, and, roughly speaking, to store the absolute value instead
of a two's complement in bits #1..15.

We may need delta + zigzag to handle these cases:

1) In front coding, we store the delta-length of the common prefix with
the previous string.  If the delta doesn't fit into one byte, which is
into the range [-127,127], we can devise a scheme to encode bigger
lengths with two bytes (this is what I've been doing in ../frenc-mini).
However, we can encode just uint16_t lengths, and pass them to dzag +
pfor16.  They are then compressed down to about 6 bits per integer.

2) In pkgelfsym, we encode .sym0 (intersection with topsym.list) with
delta + pfor16.  However, we also need to know which files reference
.sym0 and .sym1.  We assume that files are numbered 0..UINT16_MAX, in
alphabetical order:

$ head 0ad.nf
0       /usr/bin/pyrogenesis
1       /usr/lib64/0ad/libAtlasUI.so
2       /usr/lib64/0ad/libCollada.so
3       /usr/lib64/0ad/libmozjs38-ps-release.so
4       /usr/lib64/0ad/libnvcore.so
5       /usr/lib64/0ad/libnvimage.so
6       /usr/lib64/0ad/libnvmath.so
7       /usr/lib64/0ad/libnvtt.so

Symbols in .sym0 and .sym1 and unique, but file references may not be
unique (i.e. a symbol may be referenced by a few files).  Thus we will
also need to encode how many files reference each symbol.  But for now
we concentrate on the list of filenos paired with symbols, as if .sym0
and .sym1 were not unique:

$ sort -k2 0ad.f+sym0 |head -20
/usr/bin/pyrogenesis    FreeUPNPUrls
/usr/lib64/0ad/libmozjs38-ps-release.so PR_CreateThread
/usr/lib64/0ad/libmozjs38-ps-release.so PR_DestroyCondVar
/usr/lib64/0ad/libmozjs38-ps-release.so PR_DestroyLock
/usr/lib64/0ad/libmozjs38-ps-release.so PR_GetCurrentThread
/usr/lib64/0ad/libmozjs38-ps-release.so PR_JoinThread
/usr/lib64/0ad/libmozjs38-ps-release.so PR_Lock
/usr/lib64/0ad/libmozjs38-ps-release.so PR_MicrosecondsToInterval
/usr/lib64/0ad/libmozjs38-ps-release.so PR_MillisecondsToInterval
/usr/lib64/0ad/libmozjs38-ps-release.so PR_NewCondVar
/usr/lib64/0ad/libmozjs38-ps-release.so PR_NewLock
/usr/lib64/0ad/libmozjs38-ps-release.so PR_NotifyAllCondVar
/usr/lib64/0ad/libmozjs38-ps-release.so PR_NotifyCondVar
/usr/lib64/0ad/libmozjs38-ps-release.so PR_Unlock
/usr/lib64/0ad/libmozjs38-ps-release.so PR_WaitCondVar
/usr/bin/pyrogenesis    SDL_CalculateGammaRamp
/usr/bin/pyrogenesis    SDL_CreateColorCursor
/usr/bin/pyrogenesis    SDL_CreateRGBSurface
/usr/bin/pyrogenesis    SDL_CreateRGBSurfaceFrom
/usr/bin/pyrogenesis    SDL_CreateSemaphore
$ sort -k2 0ad.f+sym1 |head -20
/usr/lib64/0ad/libAtlasUI.so    Atlas_DisplayError
/usr/lib64/0ad/libAtlasUI.so    Atlas_GLSetCurrent
/usr/lib64/0ad/libAtlasUI.so    Atlas_GLSwapBuffers
/usr/lib64/0ad/libAtlasUI.so    Atlas_NotifyEndOfFrame
/usr/lib64/0ad/libAtlasUI.so    Atlas_SetConfigDirectory
/usr/lib64/0ad/libAtlasUI.so    Atlas_SetDataDirectory
/usr/lib64/0ad/libAtlasUI.so    Atlas_SetMessagePasser
/usr/lib64/0ad/libAtlasUI.so    Atlas_StartWindow
/usr/lib64/0ad/libmozjs38-ps-release.so JSID_EMPTY
/usr/lib64/0ad/libmozjs38-ps-release.so JSID_EMPTYHANDLE
/usr/lib64/0ad/libmozjs38-ps-release.so JSID_VOID
/usr/lib64/0ad/libmozjs38-ps-release.so JSID_VOIDHANDLE
/usr/lib64/0ad/libmozjs38-ps-release.so JSVAL_FALSE
/usr/bin/pyrogenesis    JSVAL_NULL
/usr/lib64/0ad/libmozjs38-ps-release.so JSVAL_NULL
/usr/lib64/0ad/libmozjs38-ps-release.so JSVAL_ONE
/usr/lib64/0ad/libmozjs38-ps-release.so JSVAL_TRUE
/usr/bin/pyrogenesis    JSVAL_VOID
/usr/lib64/0ad/libmozjs38-ps-release.so JSVAL_VOID
/usr/bin/pyrogenesis    JSVAL_ZERO

Both .sym0 and .sym1 (the second column) are encoded in alphabetical
order.  We join on filenames to encode filenos:

$ head -20 0ad.f0 |xargs echo
0 3 3 3 3 3 3 3 3 3 3 3 3 3 3 0 0 0 0 0
$ head -20 0ad.f1 |xargs echo
1 1 1 1 1 1 1 1 3 3 3 3 3 0 3 3 3 0 3 0

If we use bitpacking, then the range 0..7 can be encoded with 3 bits
per integer.  With dzag + pfor16, we get:

$ wc -l 0ad.f0 0ad.f1 |tail -1
 5463 total
$ cat 0ad.f0 0ad.f1 |pfor16-dzenc |wc -c
1614

which is 2.36 bits per integer.  Let's say we save 1 bit per integer,
and dzdec costs us 0.5 cycles per integer.  Then it must cost us 4 cycles
to save one byte.  This is close to the threshold that we use in enc64.c.
So it's not too bad already.

Now for coreutils:

$ tail coreutils.nf
101     /usr/bin/tsort
102     /usr/bin/tty
103     /usr/bin/unexpand
104     /usr/bin/uniq
105     /usr/bin/users
106     /usr/bin/vdir
107     /usr/bin/who
108     /usr/bin/whoami
109     /usr/bin/yes
110     /usr/lib/coreutils/libstdbuf.so
$ sort -k2 coreutils.f+sym0 |head
/bin/df __asprintf_chk
/usr/bin/seq    __asprintf_chk
/usr/bin/stdbuf __asprintf_chk
/usr/bin/who    __asprintf_chk
/bin/chgrp      __assert_fail
/bin/chmod      __assert_fail
/bin/chown      __assert_fail
/bin/cp __assert_fail
/bin/cut        __assert_fail
/bin/dd __assert_fail

Many files reference the same symbols.

$ head coreutils.f0
9
83
93
107
2
3
4
5
6
8

__asprintf_chk is referenced by four files (the last one being /usr/bin/who),
and then goes __assert_fail.

$ wc -l coreutils.f0 coreutils.f1 |tail -1
 7661 total
$ cat coreutils.f0 coreutils.f1 |pfor16-dzenc |wc -c
4740

With bitpacking, we would need 7 bits per integer, but with dzag +
pfor16, only 4.95.  That's already interesting enough!